split reflection and serialization test
