+++ engine/Tests/CMakeLists.txt
@@ -40,9 +40,7 @@ add_test(NAME new_rendergraph_test COMMAND new_rendergraph_test)
 set_target_properties(new_rendergraph_test PROPERTIES FOLDER engine_tests)
 
 add_subdirectory(reflection_test)
-set_target_properties(reflection_test PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/engine/Tests")
 add_subdirectory(serialization_test)
-set_target_properties(serialization_test PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/engine/Tests")
 add_subdirectory(shader_compile_test)
 set_target_properties(shader_compile_test PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/engine/Tests")
